mcp_server/requirements.txt
fastapi
uvicorn
pydantic-settings
google-generativeai
